refactor(monorepo): extract @gracechords/core + @gracechords/tokens via npm workspaces
Introduce npm workspaces and move platform-agnostic logic into packages/core
and design tokens into packages/tokens, in preparation for a shared Expo iOS
app. The web app stays at the repo root and builds/deploys unchanged.

- workspaces: root package.json "workspaces": ["packages/*"]
- packages/core (@gracechords/core): ChordPro parser/serialize/lint/convert/
  lexer/solfege/types, transposition, chord placement, instrumental, verseRef,
  songMetadata, sort, setlist codec, role hierarchy, and a Supabase factory
  (createGcSupabase). Consumed as source (no build step) via the
  @gracechords/core alias in vite.config.js plus the workspace symlink.
- compatibility shims left at every original src/... path so web imports are
  unchanged; the ChordPro renderer (pdf_mvp/media) stays web-side
- packages/tokens (@gracechords/tokens): tokens.css moved here, imported via
  src/styles/index.css
- supabase.js now wraps the core factory (injects Vite env + cookieStorage);
  sets.js delegates pure set shaping to core/setlists/setStore
- docs: AGENTS.md updated for moved paths; MONOREPO_MIGRATION.md documents the
  remaining Cloudflare-dashboard steps (Build System V2, install/Functions
  spike, apps/web move, CF cutover, watch paths)

Tests 157/157 pass; vite build emits dist/ unchanged.
