+++ packages/core/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@gracechords/core",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "description": "Platform-agnostic core: ChordPro parser, transposition, song/setlist logic, roles. No DOM or bundler deps.",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./*": "./src/*"
+  },
+  "dependencies": {
+    "@supabase/supabase-js": "^2.98.0"
+  }
+}
