+++ packages/tokens/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@gracechords/tokens",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Canonical GraceChords design tokens. Web imports tokens.css; a JS/TS token map for React Native is future work.",
+  "style": "./tokens.css",
+  "exports": {
+    "./tokens.css": "./tokens.css"
+  }
+}
