+++ src/styles/index.css
@@ -1,4 +1,4 @@
-@import './tokens.css';
+@import '../../packages/tokens/tokens.css';
 @import './fonts.css';
 @import './cards.css';
 @import './base.css';
