+++ src/utils/songs/sort.js
@@ -1,15 +1,3 @@
-export function normalizeTitleForSort(title = ''){
-  const t = String(title || '').trim()
-  const t2 = t.replace(/^[^A-Za-z0-9]+/, '') || t
-  return t2
-}
-
-export function compareSongsByTitle(a, b){
-  const aa = normalizeTitleForSort(a?.title || '')
-  const bb = normalizeTitleForSort(b?.title || '')
-  const aNum = /^[0-9]/.test(aa)
-  const bNum = /^[0-9]/.test(bb)
-  if (aNum !== bNum) return aNum ? -1 : 1
-  return aa.localeCompare(bb, undefined, { sensitivity: 'base' })
-}
-
+// Compatibility shim — re-exports from @gracechords/core. Do not add logic here.
+// The implementation moved to packages/core/src/songs/sort.
+export * from '@gracechords/core/songs/sort'
